+++ Moomie2/app/src/main/java/com/example/brianberg/moomie2/MovieObject.java
@@ -1,6 +1,19 @@
 package com.example.brianberg.moomie2;
 
 
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
+import android.os.Bundle;
+import android.util.Log;
+import android.widget.ImageView;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.net.HttpURLConnection;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.net.URLConnection;
+
 /**
  * Created by Jerry Sommerfeld on 3/29/2016.
  * A java class dedicated to the Moomie Movie database and rating system
@@ -15,7 +28,9 @@ public class MovieObject {
     String _plot;
     String _posterURL;
 
-    int moomieRating;
+    Bitmap _posterBitMap;
+
+    private int moomieRating;
 
     // Empty Constructor
     public MovieObject() {
@@ -35,6 +50,9 @@ public class MovieObject {
 
         //Initialize rating to 0
         this.moomieRating = 0;
+
+        //Initialize posterBitmap
+        setMoviePoster(_posterURL);
     }
 
     // Get the movieID
@@ -115,6 +133,7 @@ public class MovieObject {
     // Set the movie poster URL
     public void setPosterURL(String url) {
         this._posterURL = url;
+        setMoviePoster(_posterURL);
     }
 
     // get the Moomie rating
@@ -127,4 +146,80 @@ public class MovieObject {
         this.moomieRating = rating;
     }
 
+    public Bitmap get_posterBitMap() {
+        return _posterBitMap;
+    }
+
+    public void set_posterBitMap(Bitmap _posterBitMap) {
+        this._posterBitMap = _posterBitMap;
+    }
+
+    /* Mostly from http://www.tutorialspoint.com/android/android_network_connection.htm
+     * Takes a URL and returns a bitmap
+     */
+    void setMoviePoster(final String url){
+        new Thread() {
+            public void run() {
+                InputStream in = null;
+
+                //msg = Message.obtain();
+                //msg.what = 1;
+
+                try {
+                    Log.d("MovieObject: ", url);
+                    in = openHttpConnection(url);
+                    _posterBitMap = BitmapFactory.decodeStream(in);
+
+                    // ImageView ImageTooSwitch = (ImageView) view.findViewById(R.id.ImageToSwitch);
+                    //Bundle b = new Bundle();
+                    //b.putParcelable("bitmap", bitmap);
+                    //msg.setData(b);
+                    in.close();
+                }
+                catch (IOException e1) {
+                    e1.printStackTrace();
+                }
+                //messageHandler.sendMessage(msg);
+            }
+        }.start();
+
+        try{
+            Thread.sleep(300);
+        }catch (InterruptedException e){
+            Log.d("ToolsFragment","InterruptedException");
+        }
+
+        if(_posterBitMap == null) Log.d("ToolsFragment:", "Bitmap == null");
+        Log.d("ToolsFragment", "Bitmap worked");
+    }
+
+
+    // This for getting pictures, its from http://www.tutorialspoint.com/android/android_network_connection.htm
+    private InputStream openHttpConnection(String urlString){
+        InputStream in = null;
+        int resCode = -1;
+
+        try {
+            URL url = new URL(urlString);
+            URLConnection urlConn = url.openConnection();
+            if(!(urlConn instanceof HttpURLConnection)) {
+                throw new IOException("Url is not an Http URL");
+            }
+            HttpURLConnection httpConn = (HttpURLConnection) urlConn;
+            httpConn.setAllowUserInteraction(false);
+            httpConn.setInstanceFollowRedirects(true);
+            resCode = httpConn.getResponseCode();
+
+            if (resCode == HttpURLConnection.HTTP_OK) {
+                in = httpConn.getInputStream();
+            }
+        }
+        catch (MalformedURLException e) {
+            e.printStackTrace();
+        }
+        catch (IOException e) {
+            e.printStackTrace();
+        }
+        return in;
+    }
 }
