+++ Moomie2/app/src/main/java/com/example/brianberg/moomie2/ToolsFragment.java
@@ -43,49 +43,10 @@ public class ToolsFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        // Switch the image with url
+
         View view = inflater.inflate(R.layout.fragment_tools, container, false);
-        try {
-            TextView text = (TextView) view.findViewById(R.id.SaltText);
-            text.setText("Hello World");
-            ImageView ImageToSwitch = (ImageView) view.findViewById(R.id.ImageToSwitch);
-            Log.d("toolsFragmentStart: ", "ImageView Found");
-            ImageToSwitch.setImageResource(R.drawable.cow);
-            Log.d("toolsFragmentStart: ", "Changed Image");
-        }
-        catch(Exception e){
-            Log.d("Tools Fragment", "NullPointerException");
-        }
 
-        // Get Drawable from URL
-        /*try{
-            Log.d("ToolsFragment", "Getting Movie from URL");
-            InputStream in = openHttpConnection(url);
-
-            /*URL url = new URL("https://s-media-cache-ak0.pinimg.com/736x/66/b4/1c/66b41c8b9f57e17150eba935791ca4ef.jpg");
-            Bitmap bmp = BitmapFactory.decodeStream(url.openConnection().getInputStream());
-            ImageView ImageTooSwitch = (ImageView) view.findViewById(R.id.ImageToSwitch);
-            ImageTooSwitch.setImageBitmap(bmp);*/
-            /*URL url = new URL("http://ia.media-imdb.com/images/M/MV5BMTk3NDE2NzI4NF5BMl5BanBnXkFtZTgwNzE1MzEyMTE@._V1_SX300.jpg");
-            InputStream is = (InputStream) url.getContent();
-            Log.d("ToolsFragment", "got InputStream");
-
-            Drawable d = Drawable.createFromStream(is,"http://ia.media-imdb.com/images/M/MV5BMTk3NDE2NzI4NF5BMl5BanBnXkFtZTgwNzE1MzEyMTE@._V1_SX300.jpg");
-            Log.d("ToolsFragment", "got drawable");
-
-            ImageView ImageTooSwitch = (ImageView) view.findViewById(R.id.ImageToSwitch);
-            ImageTooSwitch.setImageDrawable(d);
-            Log.d("ToolsFragment", "URL process working");
-        }
-        catch (MalformedURLException e) {
-        e.printStackTrace();
-        } catch (IOException e) {
-            Log.d("ToolsFragment", "Error URL I/O Process");
-            e.printStackTrace();
-        } catch (Exception e) {
-            Log.d("ToolsFragment", "Error URL Process");
-            e.printStackTrace();
-        }*/
+        // Switch the image with url
         new Thread() {
             public void run() {
                 InputStream in = null;
@@ -125,6 +86,7 @@ public class ToolsFragment extends Fragment {
         return view;
     }
 
+    // this code is from some website http://www.tutorialspoint.com/android/android_network_connection.htm
     private InputStream openHttpConnection(String urlString){
         InputStream in = null;
         int resCode = -1;
@@ -144,11 +106,9 @@ public class ToolsFragment extends Fragment {
                 in = httpConn.getInputStream();
             }
         }
-
         catch (MalformedURLException e) {
             e.printStackTrace();
         }
-
         catch (IOException e) {
             e.printStackTrace();
         }
